Fix sorting pokemon storage
